Use pipenv for script requirements
This switches to using pipenv to manage the requirements for the vdp_scanner.py
script. This will help ensure deterministic Docker builds by installing
specific package versions, but stores them in a maintainable matter.

There was also a move back to using a root user because of issues with output
to the host machine and running under a different UID inside the Docker
container.

A debug option was added to the script and the version of chromium-common was
updated to resolve a build issue.