Switch to an older version of the chromium package
Per https://packages.debian.org/buster/chromium there is no
`90.0.4430.212-1~deb10u1` release available for the `arm64` platform. As a
result we much use the older `89.0.4389.114-1~deb10u1` version for now. I
have not been able to determine why that is the only platform out of the four
for the package that has an older build. The next time dependencies are updated
this will hopefully not be an issue.