Update and lock the Pipfile
With a properly formatted Pipfile we have no reason to keep a
requirements.txt file so it is removed. The Pipfile has been updated to
better reflect what the Docker image should have installed. The version
of cisagov/hash-http-content is manually updated and all other
dependencies are updated with `pipenv lock`.