Add support for the linux/armv7 platform
The latest version of the chromium package on Alpine 3.17 has support
for armv7 in addition to amd64 and arm64. Since that package is why we
reduced support to just amd64 and arm64 it makes sense to add support
for build armv7 images. Due to build dependencies we use the
`alpine:3.17` image for the compile-stage instead of a Python Docker
image. Additional packages were added to support building the lxml
Python package for armv7 since pre-built wheels are not available.