Update the version of Python used in the Docker image
Python 3.11.8 is no longer available for Alpine Linux 3.18. We update
the system package version installed in `compile-stage` to the version
of Python currently available for Alpine Linux 3.18. However, there is
no equivalent `python` Docker image available for Alpine Linux 3.18. As
a result we must upgrade to Alpine Linux 3.19, which fortunately has
the same version of Python available as a system package.