Use the system package version of the `lxml` Python package
I am seeing problems building a wheel for the `lxml` package when the
Docker image is build for armv7 in GitHub Actions. It appears to be
failing with the following error:
virtual memory exhausted: Out of memory
Compile failed: command '/usr/bin/gcc' failed with exit code 1

Since there is no way for us to increase the memory on the free GitHub
Action runners we currently rely on this is the sanest (to me)
workaround. The system packages that are installed to build the `lxml`
package on platforms that do not have a wheel available from PyPI are
removed as part of this change.