Bump Alpine Linux from 3.19 to 3.21
The two Docker images are updated and this includes updating the system
package versions to match the versions available for Alpine Linux 3.21.
The version of Python used in the Docker image is updated to match the
version available for Alpine Linux 3.21. The Pipfile is updated to
reflect the new versions of Python and the lxml package.